pyproject.toml
[tool.poetry]
authors = ["Iaroslav Russkikh"]
description = "Collection of scrapy spiders"
name = "agblox"
readme = "README.md"
version = "0.52.15"
repository = "https://github.com/agblox/scrapy-spiders"
documentation = "https://github.com/agblox/scrapy-spiders"
homepage = "https://github.com/agblox/scrapy-spiders"

[tool.poetry.dependencies]
python = "<3.9,>=3.8"

# dev deps
pytest = { version = "^6.1.1", optional = true }
vcrpy = { version = "^4.1.0", optional = true }

# package deps
Scrapy = "^2.3.0"
requests = "^2.24.0"
attrs = "^20.2.0"
boto3 = "^1.15.15"
PyYAML = "^5.3.1"
tweepy = "^3.9.0"
arrow = "^0.17.0"
python-dotenv = "^0.14.0"
yfinance = "^0.1.55"
watchtower = "^1.0.0"
praw = "7.1.0"
scrapy-rotating-proxies = "^0.6.2"
scrapy-selenium = {git = "https://github.com/arkadybag/scrapy-selenium.git", rev = "develop"}
feedparser = "^6.0.8"
langdetect = "^1.0.9"
pytube = "11.0.2"
pandas = "^1.3.4"
linkedin-api = "~=2.0.0a"
fredapi = "^0.4.3"
xlrd = "^2.0.1"
openpyxl = "^3.0.9"
pytest-vcr = "^1.0.2"
polygon-api-client = "^0.2.11"
intrinio-sdk = "^5.17.0"
spidermon = "^1.16.2"
schematics = "^2.1.1"
cryptography = "3.4.6"

[tool.poetry.extras]
dev = ["pytest", "vcrpy"]

[tool.poetry.urls]
issues = "https://github.com/agblox/scrapy-spiders"

[tool.black]
line-length = 100
target-version = ['py38']

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=1.1.2"]

[tool.semantic_release]
version_variable = [
    "pyproject.toml:version"
]
upload_to_pypi = false
upload_to_release = false
commit_subject = "chore(release): {version} [skip actions]"
